requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
google-genai
aiofiles==23.2.1
markdown==3.5.2
bleach==6.0.0